Move files to containers directory
Move Days.js to containers directory and replace code in User.js file with code from MainPage.js file that is now deleted
